refactor(heroes-radar): move onChange logic to parent component
